october push

+++ skyhotel-service/src/main/java/com/skybooking/skyhotelservice/v1_0_0/ui/model/response/payment/PaymentTransactionRQ.java
@@ -31,8 +31,8 @@ public class PaymentTransactionRQ {
     private String ipay88PaymentId;
     private String paymentCode;
     private String email;
-    private Integer skyuserId;
-    private Integer companyId;
+    private Long skyuserId;
+    private Long companyId;
     private String lang;
     private BookingDetailRS bookingDetail;
